+++ 06-airflow-dag-example/docker-compose.yml
@@ -5,7 +5,6 @@ x-airflow-common:
   # Comment the image line, place your Dockerfile in the directory where you placed the docker-compose.yaml
   # and uncomment the "build" line below, Then run `docker-compose build` to build the images.
   image: ${AIRFLOW_IMAGE_NAME:-apache/airflow:2.5.0}
-  # build: .
   environment:
     &airflow-common-env
     AIRFLOW_CONN_LAKEFS: '{
@@ -25,8 +24,12 @@ x-airflow-common:
     AIRFLOW__CORE__FERNET_KEY: ''
     AIRFLOW__CORE__DAGS_ARE_PAUSED_AT_CREATION: 'true'
     AIRFLOW__API__AUTH_BACKENDS: 'airflow.api.auth.backend.basic_auth'
-    _PIP_ADDITIONAL_REQUIREMENTS: "airflow-provider-lakefs apache-airflow[pandas] ${_PIP_ADDITIONAL_REQUIREMENTS:-}"
+    _PIP_ADDITIONAL_REQUIREMENTS: "dulwich airflow-provider-lakefs apache-airflow[pandas] ${_PIP_ADDITIONAL_REQUIREMENTS:-}"
+    DAG_CODE_GIT_ROOT: /opt/lakefs_example
+    EXTRACT_CODE_GIT_ROOT: /opt/lakefs_example
+    TRANSFORM_CODE_GIT_ROOT: /opt/lakefs_example
   volumes:
+    - .:/opt/lakefs_example
     - ./dags:/opt/airflow/dags
     - ./app_data/airflow/logs:/opt/airflow/logs
     - ./app_data/airflow/plugins:/opt/airflow/plugins
